+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased] 
+
+### Changed
+- AuthenticationResponseValidator.constructAuthenticationIdentity() converted into a static method
+
 ## [2.1.1] - 2021-09-06
 
 ### Fixed